approve and delete btns work
