+++ components/DeleteDialog.vue
@@ -1,0 +1,36 @@
+<template>
+  <v-dialog :value="showDeleteDialog" max-width="500px">
+    <v-card>
+      <v-card-title class="text-h5"
+        >Are you sure you want to delete ID: {{ idToDelete }}?</v-card-title
+      >
+      <v-card-actions>
+        <v-spacer></v-spacer>
+        <v-btn color="blue darken-1" text @click="handleCancelBtnPressed"
+          >Cancel</v-btn
+        >
+        <v-btn color="blue darken-1" text @click="handleConfirmBtnPressed"
+          >Delete</v-btn
+        >
+        <v-spacer></v-spacer>
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
+</template>
+
+<script>
+export default {
+  props: {
+    idToDelete: String,
+    showDeleteDialog: Boolean,
+  },
+  methods: {
+    handleCancelBtnPressed() {
+      this.$emit("on-cancel-btn-pressed");
+    },
+    handleConfirmBtnPressed() {
+      this.$emit("on-confirm-delete-btn-pressed");
+    },
+  },
+};
+</script>
